Use connection state mixin
